+++ skforecast/direct/_forecaster_direct.py
@@ -1100,6 +1100,7 @@ class ForecasterDirect(ForecasterBase):
             If `True`, in-sample residuals will be stored in the forecaster object
             after fitting (`in_sample_residuals_` and `in_sample_residuals_by_bin_`
             attributes).
+            If `False`, only the intervals of the bins are stored.
         random_state : int, default 123
             Set a seed for the random generator so that the stored sample 
             residuals are always deterministic.
@@ -1156,6 +1157,7 @@ class ForecasterDirect(ForecasterBase):
                 If `True`, in-sample residuals will be stored in the forecaster object
                 after fitting (`in_sample_residuals_` and `in_sample_residuals_by_bin_`
                 attributes).
+                If `False`, only the intervals of the bins are stored.
             random_state : int, default 123
                 Set a seed for the random generator so that the stored sample 
                 residuals are always deterministic.
@@ -2404,6 +2406,137 @@ class ForecasterDirect(ForecasterBase):
         if self.differentiation is not None:
             self.window_size += self.differentiation   
             self.differentiator.set_params(window_size=self.window_size)
+
+    def set_in_sample_residuals(
+        self,
+        y: pd.Series,
+        exog: pd.Series | pd.DataFrame | None = None,
+        random_state: int = 123
+    ) -> None:
+        """
+        Set in-sample residuals in case they were not calculated during the
+        training process. 
+        
+        In-sample residuals are calculated as the difference between the true 
+        values and the predictions made by the forecaster using the training 
+        data. The following internal attributes are updated:
+
+        + `in_sample_residuals_`: Dictionary containing a numpy ndarray with the
+        residuals for each step in the form `{step: residuals}`.
+        + `binner_intervals_`: intervals used to bin the residuals are calculated
+        using the quantiles of the predicted values.
+        + `in_sample_residuals_by_bin_`: residuals are binned according to the
+        predicted value they are associated with and stored in a dictionary, where
+        the keys are the intervals of the predicted values and the values are
+        the residuals associated with that range. 
+
+        A total of 10_000 residuals are stored in the attribute `in_sample_residuals_`.
+        If the number of residuals is greater than 10_000, a random sample of
+        10_000 residuals is stored. The number of residuals stored per bin is
+        limited to `10_000 // self.binner.n_bins_`.
+        
+        Parameters
+        ----------
+        y : pandas Series
+            Training time series.
+        exog : pandas Series, pandas DataFrame, default None
+            Exogenous variable/s included as predictor/s. Must have the same
+            number of observations as `y` and their indexes must be aligned so
+            that y[i] is regressed on exog[i].
+        random_state : int, default 123
+            Sets a seed to the random sampling for reproducible output.
+
+        Returns
+        -------
+        None
+
+        """
+
+        if not self.is_fitted:
+            raise NotFittedError(
+                "This forecaster is not fitted yet. Call `fit` with appropriate "
+                "arguments before using `set_in_sample_residuals()`."
+            )
+        
+        check_y(y=y)
+        y_index_range = preprocess_y(
+            y=y, return_values=False, suppress_warnings=True
+        )[1][[0, -1]]
+        if not y_index_range.equals(self.training_range_):
+            raise IndexError(
+                f"The index range of `y` does not match the range "
+                f"used during training. Please ensure the index is aligned "
+                f"with the training data.\n"
+                f"    Expected : {self.training_range_}\n"
+                f"    Received : {y_index_range}"
+            )
+        
+        # NOTE: This attributes are modified in _create_train_X_y, store original values
+        original_exog_in_ = self.exog_in_
+        original_X_train_window_features_names_out_ = self.X_train_window_features_names_out_
+        original_X_train_direct_exog_names_out_ = self.X_train_direct_exog_names_out_
+        
+        (
+            X_train,
+            y_train,
+            _,
+            _,
+            X_train_features_names_out_,
+            *_
+        ) = self._create_train_X_y(y=y, exog=exog)
+            
+        if not X_train_features_names_out_ == self.X_train_features_names_out_:
+
+            # NOTE: Reset attributes modified in _create_train_X_y to their original values
+            self.exog_in_ = original_exog_in_
+            self.X_train_window_features_names_out_ = original_X_train_window_features_names_out_
+            self.X_train_direct_exog_names_out_ = original_X_train_direct_exog_names_out_
+
+            raise ValueError(
+                f"Feature mismatch detected after matrix creation. The features "
+                f"generated from the provided data do not match those used during "
+                f"the training process. To correctly set in-sample residuals, "
+                f"ensure that the same data and preprocessing steps are applied.\n"
+                f"    Expected output : {self.X_train_features_names_out_}\n"
+                f"    Current output  : {X_train_features_names_out_}"
+            )
+        
+        y_true_steps = []
+        y_pred_steps = []
+        self.in_sample_residuals_ = {}
+        for step in range(1, self.steps + 1):
+
+            X_train_step, y_train_step = self.filter_train_X_y_for_step(
+                                             step          = step,
+                                             X_train       = X_train,
+                                             y_train       = y_train,
+                                             remove_suffix = True
+                                         )
+            
+            y_true_step = y_train_step.to_numpy()
+            y_pred_step = self.regressors_[step].predict(X_train_step)
+            residuals = y_true_step - y_pred_step
+            if len(residuals) > 10_000:
+                rng = np.random.default_rng(seed=random_state)
+                residuals = residuals[
+                    rng.integers(low=0, high=len(residuals), size=10_000)
+                ]
+
+            y_true_steps.append(y_true_step)
+            y_pred_steps.append(y_pred_step)
+            self.in_sample_residuals_[step] = residuals
+
+        self._binning_in_sample_residuals(
+            y_true                    = np.concatenate(y_true_steps),
+            y_pred                    = np.concatenate(y_pred_steps),
+            store_in_sample_residuals = True,
+            random_state              = random_state
+        )
+
+        # NOTE: Reset attributes modified in _create_train_X_y to their original values
+        self.exog_in_ = original_exog_in_
+        self.X_train_window_features_names_out_ = original_X_train_window_features_names_out_
+        self.X_train_direct_exog_names_out_ = original_X_train_direct_exog_names_out_
 
     def set_out_sample_residuals(
         self,