+++ skforecast/direct/tests/tests_forecaster_direct_multivariate/test_predict_dist.py
@@ -37,7 +37,7 @@ def test_predict_dist_TypeError_when_distribution_object_is_not_valid():
                      transformer_series = StandardScaler(),
                      transformer_exog   = transformer_exog
                  )
-    forecaster.fit(series=series, exog=exog)
+    forecaster.fit(series=series, exog=exog, store_in_sample_residuals=True)
     
     class CustomObject:  # pragma: no cover
         pass
@@ -71,7 +71,7 @@ def test_predict_dist_output_when_forecaster_is_LinearRegression_steps_is_2_in_s
                      transformer_exog   = transformer_exog
                  )
     
-    forecaster.fit(series=series, exog=exog)
+    forecaster.fit(series=series, exog=exog, store_in_sample_residuals=True)
     results = forecaster.predict_dist(
                   steps                   = 2,
                   exog                    = exog_predict,
@@ -106,7 +106,7 @@ def test_predict_dist_output_when_forecaster_is_LinearRegression_steps_is_2_in_s
                      transformer_exog   = transformer_exog
                  )
     
-    forecaster.fit(series=series, exog=exog)
+    forecaster.fit(series=series, exog=exog, store_in_sample_residuals=True)
     forecaster.out_sample_residuals_ = forecaster.in_sample_residuals_
     results = forecaster.predict_dist(
                   steps                   = 2,